+++ get.go
@@ -6,7 +6,6 @@ import (
 	"github.com/giantswarm/microerror"
 	"github.com/prometheus/client_golang/prometheus"
 	"helm.sh/helm/v3/pkg/action"
-	"helm.sh/helm/v3/pkg/release"
 )
 
 // GetReleaseContent gets the current status of the Helm Release including any
@@ -41,12 +40,4 @@ func (c *Client) getReleaseContent(ctx context.Context, namespace, releaseName s
 	}
 
 	return releaseToReleaseContent(res), nil
-}
-
-func releaseToReleaseContent(res *release.Release) *ReleaseContent {
-	return &ReleaseContent{
-		Name:   res.Name,
-		Status: res.Info.Status.String(),
-		Values: res.Config,
-	}
 }
